poprawienie dto, usuniecie niepotrzebnych starych dto
