+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/controller/AuthController.java
@@ -4,7 +4,6 @@ import jakarta.servlet.http.Cookie;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
-import org.apache.coyote.Response;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -22,7 +21,6 @@ import pl.uniwersytetkaliski.studenteventsplatform.model.UserRole;
 import pl.uniwersytetkaliski.studenteventsplatform.service.UserService;
 
 import java.time.LocalDateTime;
-import java.util.Base64;
 import java.util.Objects;
 
 @RestController
