--- src/main/java/pl/uniwersytetkaliski/studenteventsplatform/dto/CategoryDTO.java
@@ -1,22 +1,0 @@
-package pl.uniwersytetkaliski.studenteventsplatform.dto;
-
-public class CategoryDTO {
-    private Long id;
-    private String name;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-}
