--- src/main/java/pl/uniwersytetkaliski/studenteventsplatform/dto/EventDTO.java
@@ -1,135 +1,0 @@
-package pl.uniwersytetkaliski.studenteventsplatform.dto;
-
-import pl.uniwersytetkaliski.studenteventsplatform.model.EventStatus;
-
-
-import java.time.LocalDateTime;
-
-public class EventDTO {
-    private Long id;
-    private String name;
-    private LocationDTO locationDTO;
-    private EventStatus status;
-    private int maxCapacity;
-    private int currentCapacity;
-    private LocalDateTime creationDate;
-    private LocalDateTime startDate;
-    private LocalDateTime endDate;
-    private String description;
-    private CategoryDTO categoryDTO;
-    private boolean deleted;
-    private LocalDateTime deleted_at;
-    private String created_by;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public LocationDTO getLocationDTO() {
-        return locationDTO;
-    }
-
-    public void setLocationDTO(LocationDTO locationDTO) {
-        this.locationDTO = locationDTO;
-    }
-
-    public EventStatus getStatus() {
-        return status;
-    }
-
-    public void setStatus(EventStatus status) {
-        this.status = status;
-    }
-
-    public int getMaxCapacity() {
-        return maxCapacity;
-    }
-
-    public void setMaxCapacity(int maxCapacity) {
-        this.maxCapacity = maxCapacity;
-    }
-
-    public int getCurrentCapacity() {
-        return currentCapacity;
-    }
-
-    public void setCurrentCapacity(int currentCapacity) {
-        this.currentCapacity = currentCapacity;
-    }
-
-    public LocalDateTime getCreationDate() {
-        return creationDate;
-    }
-
-    public void setCreationDate(LocalDateTime creationDate) {
-        this.creationDate = creationDate;
-    }
-
-    public LocalDateTime getStartDate() {
-        return startDate;
-    }
-
-    public void setStartDate(LocalDateTime startDate) {
-        this.startDate = startDate;
-    }
-
-    public LocalDateTime getEndDate() {
-        return endDate;
-    }
-
-    public void setEndDate(LocalDateTime endDate) {
-        this.endDate = endDate;
-    }
-
-    public String getDescription() {
-        return description;
-    }
-
-    public void setDescription(String description) {
-        this.description = description;
-    }
-
-    public CategoryDTO getCategoryDTO() {
-        return categoryDTO;
-    }
-
-    public void setCategoryDTO(CategoryDTO categoryDTO) {
-        this.categoryDTO = categoryDTO;
-    }
-
-    public boolean isDeleted() {
-        return deleted;
-    }
-
-    public void setDeleted(boolean deleted) {
-        this.deleted = deleted;
-    }
-
-    public LocalDateTime getDeleted_at() {
-        return deleted_at;
-    }
-
-    public void setDeleted_at(LocalDateTime deleted_at) {
-        this.deleted_at = deleted_at;
-    }
-
-    public String getCreated_by() {
-        return created_by;
-    }
-
-    public void setCreated_by(String created_by) {
-        this.created_by = created_by;
-    }
-}
