--- src/main/java/pl/uniwersytetkaliski/studenteventsplatform/dto/LocationDTO.java
@@ -1,49 +1,0 @@
-package pl.uniwersytetkaliski.studenteventsplatform.dto;
-
-public class LocationDTO {
-    private Long id;
-    private String city;
-    private String street;
-    private String houseNumber;
-    private String postalCode;
-
-    public Long getId() {
-        return id;
-    }
-
-    public void setId(Long id) {
-        this.id = id;
-    }
-
-    public String getCity() {
-        return city;
-    }
-
-    public void setCity(String city) {
-        this.city = city;
-    }
-
-    public String getStreet() {
-        return street;
-    }
-
-    public void setStreet(String street) {
-        this.street = street;
-    }
-
-    public String getHouseNumber() {
-        return houseNumber;
-    }
-
-    public void setHouseNumber(String houseNumber) {
-        this.houseNumber = houseNumber;
-    }
-
-    public String getPostalCode() {
-        return postalCode;
-    }
-
-    public void setPostalCode(String postalCode) {
-        this.postalCode = postalCode;
-    }
-}
