+++ src/main/java/pl/uniwersytetkaliski/studenteventsplatform/dto/eventDTO/EventResponseDTO.java
@@ -15,8 +15,8 @@ public class EventResponseDTO {
     private int maxCapacity;
     private int currentCapacity;
     private LocalDateTime creationDate;
-    private LocalDateTime startDate;
-    private LocalDateTime endDate;
+    private LocalDateTime startDateTime;
+    private LocalDateTime endDateTime;
     private String description;
     private CategoryResponseDTO categoryDTO;
     private boolean deleted;
@@ -80,20 +80,20 @@ public class EventResponseDTO {
         this.creationDate = creationDate;
     }
 
-    public LocalDateTime getStartDate() {
-        return startDate;
+    public LocalDateTime getStartDateTime() {
+        return startDateTime;
     }
 
-    public void setStartDate(LocalDateTime startDate) {
-        this.startDate = startDate;
+    public void setStartDateTime(LocalDateTime startDateTime) {
+        this.startDateTime = startDateTime;
     }
 
-    public LocalDateTime getEndDate() {
-        return endDate;
+    public LocalDateTime getEndDateTime() {
+        return endDateTime;
     }
 
-    public void setEndDate(LocalDateTime endDate) {
-        this.endDate = endDate;
+    public void setEndDateTime(LocalDateTime endDateTime) {
+        this.endDateTime = endDateTime;
     }
 
     public String getDescription() {
